mtd: spi-nor: spansion: Fix support for cy15x104q devices
Cypress 4 Mbit FRAM.

These devices are divided into families where the minor byte only reflects
differences such as grade (automotive/industrial/etc.) and not FRAM size.

Introduce a private macro FRAM_INFO2 to match an entire family, because
amount of devices in these families is large. If not true, if a device in
a family is found with different size, it can be added before family
definition, using different macro, so that it will be more specific and
parsed before.

Use recently added continuation codes support to ignore 6 codes.

Signed-off-by: Roman Volkov <riv10@pm.me>